EES-221 Split out data file components from data uploads section
This includes `DataFileUploadForm`, which has been refactored so that is
extensible enough for re-use in the new data replacement pages.
