+++ src/explore-education-statistics-admin/src/pages/release/data/components/ReleaseDataUploadsSection.tsx
@@ -1,120 +1,40 @@
-import ImporterStatus from '@admin/components/ImporterStatus';
-import useFormSubmit from '@admin/hooks/useFormSubmit';
+import DataFileSummaryList from '@admin/pages/release/data/components/DataFileSummaryList';
+import DataFileUploadForm, {
+  DataFileUploadFormValues,
+} from '@admin/pages/release/data/components/DataFileUploadForm';
 import releaseDataFileService, {
   DataFile,
   DeleteDataFilePlan,
   ImportStatusCode,
 } from '@admin/services/releaseDataFileService';
-import releaseMetaFileService from '@admin/services/releaseMetaFileService';
 import Accordion from '@common/components/Accordion';
 import AccordionSection from '@common/components/AccordionSection';
-import Button from '@common/components/Button';
-import ButtonGroup from '@common/components/ButtonGroup';
 import ButtonText from '@common/components/ButtonText';
-import { Form, FormFieldRadioGroup } from '@common/components/form';
-import FormFieldFileInput from '@common/components/form/FormFieldFileInput';
 import FormFieldTextInput from '@common/components/form/FormFieldTextInput';
 import LoadingSpinner from '@common/components/LoadingSpinner';
 import ModalConfirm from '@common/components/ModalConfirm';
-import SummaryList from '@common/components/SummaryList';
 import SummaryListItem from '@common/components/SummaryListItem';
 import WarningMessage from '@common/components/WarningMessage';
 import useAsyncHandledRetry from '@common/hooks/useAsyncHandledRetry';
 import logger from '@common/services/logger';
-import {
-  FieldMessageMapper,
-  mapFieldErrors,
-} from '@common/validation/serverValidations';
+import { mapFieldErrors } from '@common/validation/serverValidations';
 import Yup from '@common/validation/yup';
-import { format } from 'date-fns';
-import { Formik } from 'formik';
-import React, { useState } from 'react';
+import React, { useCallback, useState } from 'react';
 
-interface FormValues {
+interface FormValues extends DataFileUploadFormValues {
   subjectTitle: string;
-  uploadType: 'csv' | 'zip';
-  dataFile: File | null;
-  metadataFile: File | null;
-  zipFile: File | null;
 }
 
-const errorMappings = (
-  values: FormValues,
-): FieldMessageMapper<FormValues>[] => {
-  const baseMappings = [
-    mapFieldErrors<FormValues>({
-      target: 'subjectTitle',
-      messages: {
-        SUBJECT_TITLE_MUST_BE_UNIQUE: 'Subject title must be unique',
-        SUBJECT_TITLE_CANNOT_CONTAIN_SPECIAL_CHARACTERS:
-          'Subject title cannot contain special characters',
-      },
-    }),
-  ];
-
-  if (values.uploadType === 'zip') {
-    return [
-      ...baseMappings,
-      mapFieldErrors<FormValues>({
-        target: 'zipFile',
-        messages: {
-          DATA_FILE_MUST_BE_ZIP_FILE: 'Choose a valid ZIP file',
-          DATA_ZIP_FILE_CAN_ONLY_CONTAIN_TWO_FILES:
-            'ZIP file can only contain two CSV files',
-          DATA_ZIP_FILE_DOES_NOT_CONTAIN_CSV_FILES:
-            'ZIP file does not contain any CSV files',
-          DATA_ZIP_FILE_ALREADY_EXISTS: 'ZIP file already exists',
-          CANNOT_OVERWRITE_DATA_FILE: 'Choose a unique ZIP data file name',
-          DATA_AND_METADATA_FILES_CANNOT_HAVE_THE_SAME_NAME:
-            'ZIP data and metadata filenames cannot be the same',
-          DATA_FILE_CANNOT_BE_EMPTY: 'Choose a ZIP data file that is not empty',
-          DATA_FILENAME_CANNOT_CONTAIN_SPACES_OR_SPECIAL_CHARACTERS:
-            'ZIP data filename cannot contain spaces or special characters',
-          CANNOT_OVERWRITE_METADATA_FILE:
-            'Choose a unique ZIP metadata file name',
-          METADATA_FILE_CANNOT_BE_EMPTY:
-            'Choose a ZIP metadata file that is not empty',
-          META_FILENAME_CANNOT_CONTAIN_SPACES_OR_SPECIAL_CHARACTERS:
-            'ZIP metadata filename cannot contain spaces or special characters',
-          META_FILE_IS_INCORRECTLY_NAMED:
-            'ZIP metadata filename must end with .meta.csv',
-        },
-      }),
-    ];
-  }
-
-  return [
-    ...baseMappings,
-    mapFieldErrors<FormValues>({
-      target: 'dataFile',
-      messages: {
-        CANNOT_OVERWRITE_DATA_FILE: 'Choose a unique data file name',
-        DATA_AND_METADATA_FILES_CANNOT_HAVE_THE_SAME_NAME:
-          'Choose a different file name for data and metadata files',
-        DATA_FILE_CANNOT_BE_EMPTY: 'Choose a data file that is not empty',
-        DATA_FILE_MUST_BE_CSV_FILE:
-          'Data file must be a CSV with UTF-8 encoding',
-        DATA_FILENAME_CANNOT_CONTAIN_SPACES_OR_SPECIAL_CHARACTERS:
-          'Data filename cannot contain spaces or special characters',
-        DATA_FILE_ALREADY_UPLOADED: 'Data file has already been uploaded',
-      },
-    }),
-    mapFieldErrors<FormValues>({
-      target: 'metadataFile',
-      messages: {
-        CANNOT_OVERWRITE_METADATA_FILE: 'Choose a unique metadata file name',
-        METADATA_FILE_CANNOT_BE_EMPTY:
-          'Choose a metadata file that is not empty',
-        META_FILE_MUST_BE_CSV_FILE:
-          'Metadata file must be a CSV with UTF-8 encoding',
-        META_FILENAME_CANNOT_CONTAIN_SPACES_OR_SPECIAL_CHARACTERS:
-          'Metadata filename cannot contain spaces or special characters',
-        META_FILE_IS_INCORRECTLY_NAMED:
-          'Metadata filename is incorrectly named',
-      },
-    }),
-  ];
-};
+const errorMappings = [
+  mapFieldErrors<FormValues>({
+    target: 'subjectTitle',
+    messages: {
+      SUBJECT_TITLE_MUST_BE_UNIQUE: 'Subject title must be unique',
+      SUBJECT_TITLE_CANNOT_CONTAIN_SPECIAL_CHARACTERS:
+        'Subject title cannot contain special characters',
+    },
+  }),
+];
 
 interface Props {
   releaseId: string;
@@ -130,7 +50,6 @@ const formId = 'dataFileUploadForm';
 
 const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
   const [deleteDataFile, setDeleteDataFile] = useState<DeleteDataFile>();
-  const [isUploading, setIsUploading] = useState(false);
 
   const {
     value: dataFiles = [],
@@ -174,12 +93,10 @@ const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
     });
   };
 
-  const handleSubmit = useFormSubmit<FormValues>(async (values, actions) => {
-    setIsUploading(true);
+  const handleSubmit = useCallback(
+    async (values: FormValues) => {
+      let file: DataFile;
 
-    let file: DataFile;
-
-    try {
       if (values.uploadType === 'csv') {
         file = await releaseDataFileService.uploadDataFiles(releaseId, {
           name: values.subjectTitle,
@@ -193,187 +110,86 @@ const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
         });
       }
 
-      actions.resetForm();
-
       setDataFiles({
         value: [...dataFiles, file],
       });
-    } finally {
-      setIsUploading(false);
-    }
-  }, errorMappings);
+    },
+    [dataFiles, releaseId, setDataFiles],
+  );
 
   return (
     <>
-      <Formik<FormValues>
-        enableReinitialize
-        initialValues={{
-          subjectTitle: '',
-          uploadType: 'csv',
-          dataFile: null,
-          metadataFile: null,
-          zipFile: null,
-        }}
-        onReset={() => {
-          document
-            .querySelectorAll(`#${formId} input[type='file']`)
-            .forEach(input => {
-              const fileInput = input as HTMLInputElement;
-              fileInput.value = '';
-            });
-        }}
-        onSubmit={handleSubmit}
-        validationSchema={Yup.object<FormValues>({
-          subjectTitle: Yup.string()
-            .required('Enter a subject title')
-            .test({
-              name: 'unique',
-              message: 'Enter a unique subject title',
-              test(value: string) {
-                if (!value) {
-                  return true;
-                }
+      <h2>Add data file to release</h2>
+      <div className="govuk-inset-text">
+        <h3>Before you start</h3>
 
-                return (
-                  dataFiles.find(
-                    f => f.title.toUpperCase() === value.toUpperCase(),
-                  ) === undefined
-                );
-              },
-            }),
-          uploadType: Yup.mixed<FormValues['uploadType']>().oneOf([
-            'csv',
-            'zip',
-          ]),
-          dataFile: Yup.file().when('uploadType', {
-            is: 'csv',
-            then: Yup.file()
-              .required('Choose a data file')
-              .minSize(0, 'Choose a data file that is not empty'),
-          }),
-          metadataFile: Yup.file().when('uploadType', {
-            is: 'csv',
-            then: Yup.file()
-              .required('Choose a metadata file')
-              .minSize(0, 'Choose a metadata file that is not empty'),
-          }),
-          zipFile: Yup.file().when('uploadType', {
-            is: 'zip',
-            then: Yup.file()
-              .required('Choose a zip file')
-              .minSize(0, 'Choose a ZIP file that is not empty'),
-          }),
-        })}
-      >
-        {form => (
-          <>
-            <h2>Add data file to release</h2>
+        <ul>
+          <li>
+            make sure your data files have passed the checks in our{' '}
+            <a href="https://rsconnect/rsc/dfe-published-data-qa/">
+              screening app
+            </a>
+          </li>
+          <li>
+            if your data does not meet these standards, you won’t be able to
+            upload it to your release
+          </li>
+          <li>
+            if you have any issues uploading data and files, or questions about
+            data standards contact:{' '}
+            <a href="mailto:explore.statistics@education.gov.uk">
+              explore.statistics@education.gov.uk
+            </a>
+          </li>
+        </ul>
+      </div>
+      {canUpdateRelease ? (
+        <DataFileUploadForm
+          id={formId}
+          initialValues={{
+            subjectTitle: '',
+            uploadType: 'csv',
+            dataFile: null,
+            metadataFile: null,
+            zipFile: null,
+          }}
+          errorMappings={errorMappings}
+          onSubmit={handleSubmit}
+          validationSchema={baseSchema =>
+            baseSchema.shape({
+              subjectTitle: Yup.string()
+                .required('Enter a subject title')
+                .test({
+                  name: 'unique',
+                  message: 'Enter a unique subject title',
+                  test(value: string) {
+                    if (!value) {
+                      return true;
+                    }
 
-            <div className="govuk-inset-text">
-              <h3>Before you start</h3>
-
-              <ul>
-                <li>
-                  make sure your data files have passed the checks in our{' '}
-                  <a href="https://rsconnect/rsc/dfe-published-data-qa/">
-                    screening app
-                  </a>
-                </li>
-                <li>
-                  if your data does not meet these standards, you won’t be able
-                  to upload it to your release
-                </li>
-                <li>
-                  if you have any issues uploading data and files, or questions
-                  about data standards contact:{' '}
-                  <a href="mailto:explore.statistics@education.gov.uk">
-                    explore.statistics@education.gov.uk
-                  </a>
-                </li>
-              </ul>
-            </div>
-
-            {canUpdateRelease ? (
-              <Form id={formId}>
-                <div style={{ position: 'relative' }}>
-                  {isUploading && (
-                    <LoadingSpinner text="Uploading files" overlay />
-                  )}
-
-                  <FormFieldTextInput<FormValues>
-                    id={`${formId}-subjectTitle`}
-                    name="subjectTitle"
-                    label="Subject title"
-                    width={20}
-                  />
-
-                  <FormFieldRadioGroup<FormValues>
-                    id={`${formId}-uploadType`}
-                    name="uploadType"
-                    legend="Choose upload method"
-                    options={[
-                      {
-                        label: 'CSV files',
-                        value: 'csv',
-                        conditional: (
-                          <>
-                            <FormFieldFileInput<FormValues>
-                              id={`${formId}-dataFile`}
-                              name="dataFile"
-                              label="Upload data file"
-                              accept=".csv"
-                            />
-
-                            <FormFieldFileInput<FormValues>
-                              id={`${formId}-metadataFile`}
-                              name="metadataFile"
-                              label="Upload metadata file"
-                              accept=".csv"
-                            />
-                          </>
-                        ),
-                      },
-                      {
-                        label: 'ZIP file',
-                        hint: 'Recommended for larger data files',
-                        value: 'zip',
-                        conditional: (
-                          <FormFieldFileInput<FormValues>
-                            id={`${formId}-zipFile`}
-                            hint="Must contain both the data and metadata CSV files"
-                            name="zipFile"
-                            label="Upload ZIP file"
-                            accept=".zip"
-                          />
-                        ),
-                      },
-                    ]}
-                  />
-
-                  <ButtonGroup>
-                    <Button type="submit" disabled={form.isSubmitting}>
-                      Upload data files
-                    </Button>
-
-                    <ButtonText
-                      disabled={form.isSubmitting}
-                      onClick={() => {
-                        form.resetForm();
-                      }}
-                    >
-                      Cancel
-                    </ButtonText>
-                  </ButtonGroup>
-                </div>
-              </Form>
-            ) : (
-              <WarningMessage>
-                This release has been approved, and can no longer be updated.
-              </WarningMessage>
-            )}
-          </>
-        )}
-      </Formik>
+                    return (
+                      dataFiles.find(
+                        f => f.title.toUpperCase() === value.toUpperCase(),
+                      ) === undefined
+                    );
+                  },
+                }),
+            })
+          }
+          beforeFields={
+            <FormFieldTextInput<FormValues>
+              id={`${formId}-subjectTitle`}
+              name="subjectTitle"
+              label="Subject title"
+              width={20}
+            />
+          }
+        />
+      ) : (
+        <WarningMessage>
+          This release has been approved, and can no longer be updated.
+        </WarningMessage>
+      )}
 
       <hr className="govuk-!-margin-top-6 govuk-!-margin-bottom-6" />
 
@@ -392,55 +208,11 @@ const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
                   {dataFile.isDeleting && (
                     <LoadingSpinner text="Deleting files" overlay />
                   )}
-                  <SummaryList>
-                    <SummaryListItem term="Subject title">
-                      {dataFile.title}
-                    </SummaryListItem>
-                    <SummaryListItem term="Data file">
-                      <ButtonText
-                        onClick={() =>
-                          releaseDataFileService.downloadDataFile(
-                            releaseId,
-                            dataFile.filename,
-                          )
-                        }
-                      >
-                        {dataFile.filename}
-                      </ButtonText>
-                    </SummaryListItem>
-                    <SummaryListItem term="Metadata file">
-                      <ButtonText
-                        onClick={() =>
-                          releaseMetaFileService.downloadDataMetadataFile(
-                            releaseId,
-                            dataFile.metadataFilename,
-                          )
-                        }
-                      >
-                        {dataFile.metadataFilename}
-                      </ButtonText>
-                    </SummaryListItem>
-                    <SummaryListItem term="Data file size">
-                      {dataFile.fileSize.size.toLocaleString()}{' '}
-                      {dataFile.fileSize.unit}
-                    </SummaryListItem>
-                    <SummaryListItem term="Number of rows">
-                      {dataFile.rows.toLocaleString()}
-                    </SummaryListItem>
-
-                    <ImporterStatus
-                      releaseId={releaseId}
-                      dataFile={dataFile}
-                      onStatusChangeHandler={handleStatusChange}
-                    />
-                    <SummaryListItem term="Uploaded by">
-                      <a href={`mailto:${dataFile.userName}`}>
-                        {dataFile.userName}
-                      </a>
-                    </SummaryListItem>
-                    <SummaryListItem term="Date uploaded">
-                      {format(dataFile.created, 'd MMMM yyyy HH:mm')}
-                    </SummaryListItem>
+                  <DataFileSummaryList
+                    dataFile={dataFile}
+                    releaseId={releaseId}
+                    onStatusChange={handleStatusChange}
+                  >
                     {canUpdateRelease && dataFile.canDelete && (
                       <SummaryListItem
                         term="Actions"
@@ -462,7 +234,7 @@ const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
                         }
                       />
                     )}
-                  </SummaryList>
+                  </DataFileSummaryList>
                 </div>
               </AccordionSection>
             ))}
@@ -471,7 +243,6 @@ const ReleaseDataUploadsSection = ({ releaseId, canUpdateRelease }: Props) => {
           <p className="govuk-inset-text">No data files have been uploaded.</p>
         )}
       </LoadingSpinner>
-
       {deleteDataFile && (
         <ModalConfirm
           mounted
